Add iterator functions, loops example
Noticed am issue with generators not being appropriately delimited;
subsequent calls to a generator will return to the original environment.

For example, this implementation of `iter->list` currently does not
work:

```lisp
($define! iter->list
  ($lambda (iter)
    ($define! go
      ($lambda (acc)
        ($let ((x (iter)))
          ($if (inert? x)
            (reverse acc)
            (go (cons x acc))))))
    (go ())))
```

Inside of the `$let`, after calling `(iter)`, `acc` is always `()`,
resulting in the empty list.